feat: moved where the user query is being done
